+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - JobsReborn integration missed some null checks resulting in not working actions and conditions
 - concurrency issue when loading player data on join causing multiple profiles to get created with a database exception
 - `extends` in `conversations` sometimes crashing with IllegalStateException
+- `section` placeholder did not include target path to subsections
 ### Security
 
 ## [3.2.0] - 2026-08-17